+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+
+* enhancements
+  * [Builtin] Added support to the %R sigil. The same as %r, but without interpolation or escaping.
+
 # v0.5.0 (2012-05-24)
 
 * First official release